pyproject.toml
[project]
name = "EXPERIMENT_NAME"
version = "0.1.0"
description = "EXPERIMENT_DESCRIPTION"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "pyarrow>=15.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
